Split wire contracts by responsibility
